refactor : presentation -> view로 변경, inputView, outputView 분리
