+++ src/main/kotlin/lotto/view/OutputView.kt
@@ -1,4 +1,4 @@
-package lotto.presentation
+package lotto.view
 
-class OutputView {
+object OutputView {
 }
